Join HoeffdingTreeClassifier nodes
